+++ Indicium/Properties/AssemblyInfo.cs
@@ -11,7 +11,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
 [assembly: AssemblyProduct("Indicium")]
-[assembly: AssemblyCopyright("deadmoon © 2017")]
+[assembly: AssemblyCopyright("deadmoon © 2018")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]