+++ Indicium/Utilities.cs
@@ -52,18 +52,18 @@ namespace Indicium
 
         private static string DecodeProductKey(byte[] digitalProductId)
         {
-            var key = string.Empty;
+            string key = string.Empty;
             const int keyOffset = 52;
-            var isWin8 = (byte)((digitalProductId[66] / 6) & 1);
+            byte isWin8 = (byte)((digitalProductId[66] / 6) & 1);
             digitalProductId[66] = (byte)((digitalProductId[66] & 0xf7) | (isWin8 & 2) * 4);
 
             // Possible alpha-numeric characters in product key.
             const string digits = "BCDFGHJKMPQRTVWXY2346789";
             int last = 0;
-            for (var i = 24; i >= 0; i--)
+            for (int i = 24; i >= 0; i--)
             {
-                var current = 0;
-                for (var j = 14; j >= 0; j--)
+                int current = 0;
+                for (int j = 14; j >= 0; j--)
                 {
                     current = current * 256;
                     current = digitalProductId[j + keyOffset] + current;
@@ -73,12 +73,12 @@ namespace Indicium
                 }
                 key = digits[current] + key;
             }
-            var keypart1 = key.Substring(1, last);
+            string keypart1 = key.Substring(1, last);
             const string insert = "N";
             key = key.Substring(1).Replace(keypart1, keypart1 + insert);
             if (last == 0)
                 key = insert + key;
-            for (var i = 5; i < key.Length; i += 6)
+            for (int i = 5; i < key.Length; i += 6)
             {
                 key = key.Insert(i, "-");
             }
@@ -87,24 +87,26 @@ namespace Indicium
 
         public static string GetProductKey()
         {
-            var key = RegistryKey.OpenBaseKey(RegistryHive.LocalMachine,
+            RegistryKey key = RegistryKey.OpenBaseKey(RegistryHive.LocalMachine,
                                           RegistryView.Default);
             const string keyPath = @"Software\Microsoft\Windows NT\CurrentVersion";
-            var digitalProductId = (byte[])key.OpenSubKey(keyPath).GetValue("DigitalProductId");
+            byte[] digitalProductId = (byte[])key.OpenSubKey(keyPath, false).GetValue("DigitalProductId");
 
-            var isWin8OrUp =
-                (Environment.OSVersion.Version.Major == 6 && System.Environment.OSVersion.Version.Minor >= 2)
-                ||
-                (Environment.OSVersion.Version.Major > 6);
-
-            var productKey = isWin8OrUp ? DecodeProductKey(digitalProductId) : DecodeProductKey(digitalProductId);
-
-            if (productKey.Count() > 29)
+            try
             {
-                productKey = productKey.Substring(0, 29);
-            }
+                string productKey = DecodeProductKey(digitalProductId);
 
-            return productKey;
+                if (productKey.Count() > 29)
+                {
+                    productKey = productKey.Substring(0, 29);
+                }
+
+                return productKey;
+            }
+            catch
+            {
+                return string.Empty;
+            }
         }
     }
 }